refactor: enhance contexts naming
